% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11ytables-package.R
\docType{package}
\name{a11ytables-package}
\alias{a11ytables}
\alias{a11ytables-package}
\title{a11ytables: Create Spreadsheet Publications Following Best Practice}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generate spreadsheet publications that follow best practice guidance from the UK government's Analysis Function Central Team, with a focus on accessibility. Based on {openxlsx}. See also the Python package 'gptables'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://co-analysis.github.io/a11ytables/}
  \item \url{https://github.com/co-analysis/a11ytables}
  \item Report bugs at \url{https://github.com/co-analysis/a11ytables/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

Other contributors:
\itemize{
  \item Crown Copyright [copyright holder, funder]
  \item Tim Taylor [contributor]
  \item Matt Kerlogue [contributor]
}

}
\keyword{internal}
