% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11ytables2-package.R
\docType{package}
\name{a11ytables2-package}
\alias{a11ytables2}
\alias{a11ytables2-package}
\title{a11ytables2: Create Spreadsheet Publications Following Best Practice}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generate spreadsheet publications that follow best practice guidance from the UK government's Analysis Function Central Team, with a focus on accessibility. Based on {openxlsx2}. Successor to 'a11ytables'. See also the Python package 'gptables'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/a11ytables2}
  \item Report bugs at \url{https://github.com/matt-dray/a11ytables2/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

Other contributors:
\itemize{
  \item Crown Copyright [copyright holder, funder]
  \item Tim Taylor [contributor]
  \item Matt Kerlogue [contributor]
}

}
\keyword{internal}
