% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_contents}
\alias{add_contents}
\title{Add a Table of Contents Sheet to a Workbook}
\usage{
add_contents(wb, content)
}
\arguments{
\item{wb}{A Workbook-class object created by \link[openxlsx]{createWorkbook}}

\item{content}{A data.frame-class object containing your data.}
}
\value{
A Workbook-class object.
}
\description{
After creating a new empty workbook, named tabs and a cover sheet, you need
to add a table-of-contents sheet
}
\details{
The 'content' object should adhere to the form demonstrated in
\code{\link{lfs_tables}}
}
\examples{
\dontrun{
openxlsx::createWorkbook() \%>\%
  add_tabs(content) \%>\%
  add_cover(content) \%>\%
  add_contents(content)}
}
