% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_notes}
\alias{add_notes}
\title{Add a Notes Sheet to a Workbook}
\usage{
add_notes(wb, content)
}
\arguments{
\item{wb}{A Workbook-class object created by \link[openxlsx]{createWorkbook}}

\item{content}{A data.frame-class object containing your data.}
}
\value{
A Workbook-class object.
}
\description{
After creating a new empty workbook, named tabs, a cover sheet and a
table-of-contents sheet, you can add a notes sheet (optional, only needed if
there are any notes in teh workbook).
}
\details{
The 'content' object should adhere to the form demonstrated in
\code{\link{lfs_tables}}
}
\examples{
\dontrun{openxlsx::createWorkbook |>
  add_tabs(content) |>
  add_cover(content) |>
  add_contents(content) |>
  add_notes(content)}
}
