% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_tables}
\alias{add_tables}
\title{Add a Table Sheet to a Workbook}
\usage{
add_tables(wb, content, table_name)
}
\arguments{
\item{wb}{A Workbook-class object created by \link[openxlsx]{createWorkbook}}

\item{content}{A data.frame-class object containing your data.}

\item{table_name}{Character. The name of the table, which is used to filter
the relevant data from the object in the content argument.}
}
\value{
A Workbook-class object.
}
\description{
After creating a new empty workbook, named tabs, a cover sheet, a
table-of-contents sheet and a notes sheet (optional), you should add the
tables of data.
}
\details{
The 'content' object should adhere to the form demonstrated in
\code{\link{lfs_tables}}
}
\examples{
\dontrun{
openxlsx::createWorkbook |>
  add_tabs(content) |>
  add_cover(content) |>
  add_contents(content) |>
  add_notes(content) |>
  add_tables(content, "1a")}
}
