% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_tabs}
\alias{add_tabs}
\title{Add Named Tabs to a Workbook}
\usage{
add_tabs(wb, content)
}
\arguments{
\item{wb}{A Workbook-class object created by \link[openxlsx]{createWorkbook}}

\item{content}{A data.frame-class object containing your data.}
}
\value{
A Workbook-class object.
}
\description{
After creating a new empty workbook you need to add named tabs for the
sheets.
}
\details{
The 'content' object should adhere to the form demonstrated in
\code{\link{lfs_tables}}
}
\examples{
\dontrun{openxlsx::createWorkbook() \%>\% add_tabs(content)}
}
