% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as_a11ytable}
\alias{as_a11ytable}
\title{Coerce to an a11ytable}
\usage{
as_a11ytable(x)
}
\arguments{
\item{x}{A data.frame object to coerce.}
}
\value{
An object of class a11ytable.
}
\description{
Coerce a data.frame to an a11ytable-class object
if possible.
}
