% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{at_template_a11ytable}
\alias{at_template_a11ytable}
\title{Insert 'a11ytable' Template}
\usage{
at_template_a11ytable()
}
\description{
Insert at the cursor a template for \code{\link{new_a11ytable}} from the
'a11ytable' package, pre-filled with example information.
}
