% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{at_template_df}
\alias{at_template_df}
\title{Insert Table Templates Using 'data.frame'}
\usage{
at_template_df()
}
\description{
Insert at the cursor templates for cover, contents and notes
tables, pre-filled with example information. Uses \code{\link{data.frame}},
so isn't dependent on external packages.
}
