% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{at_template_tibble}
\alias{at_template_tibble}
\title{Insert Table Templates Using 'tibble'}
\usage{
at_template_tibble()
}
\description{
Insert at the cursor templates for cover, contents and notes tables,
pre-filled with example information. Requires the 'tibble' package to be
installed.
}
