% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook.R
\name{create_a11y_wb}
\alias{create_a11y_wb}
\title{Create An Accessible Workbook}
\usage{
create_a11y_wb(a11ytable)
}
\arguments{
\item{a11ytable}{An a11ytable-class object containing the data and
information needed to create your workbook.}
}
\value{
A Workbook-class object.
}
\description{
Populate an 'openxlsx' Workbook-class object with content from an
a11ytable-class object.
}
\details{
You can create an a11ytable object with
\code{\link{create_a11ytable}} (or \code{\link{as_a11ytable}}).
}
\examples{
\dontrun{
x <- as_a11ytable(mtcars_df)
create_a11y_wb(x)
}

}
