% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is_a11ytable}
\alias{is_a11ytable}
\title{Check for a11ytable class}
\usage{
is_a11ytable(x)
}
\arguments{
\item{x}{An object to assess.}
}
\value{
Logical. \code{TRUE} if the object has class a11ytable,
otherwise \code{FALSE}.
}
\description{
Detect whether a given object has the class
a11ytable.
}
