% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfs_tables}
\alias{lfs_tables}
\title{Labour Market Overview Tables}
\format{
A data frame with 5 rows and 7 columns:
\describe{
\item{tab_title}{Character. Text to appear on the sheet's tab.}
\item{sheet_type}{Character. Either 'cover', 'contents', 'notes', or 'tables'.}
\item{sheet_title}{Character. Title that will appear in the top-left of each sheet}
\item{source}{Character. The origin of the data.}
\item{table_name}{Character. A name applied to the marked-up table.}
\item{table}{List containing dataframes/tibbles. Rectangular table containing the data.}
}
}
\source{
\url{https://gss.civilservice.gov.uk/policy-store/releasing-statistics-in-spreadsheets/}
}
\usage{
lfs_tables
}
\description{
A dataset containing selected content from the Labour Market Overview Tables
publication (UK, December 2020), published by the Office for National
Statistics as a demonstration of good practice for releasing statistics in
spreadsheets (June 2021). See \code{\link{lfs_subtables}} for a version
with sub-tables.
}
\details{
The object is a data.frame and is formatted for easy coercion to
a11ytable class with \code{\link{as_a11ytable}}.
}
\keyword{datasets}
