% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.a11ytable}
\alias{print.a11ytable}
\title{Print a11ytable Object For Reading}
\usage{
\method{print}{a11ytable}(x, ...)
}
\arguments{
\item{x}{An a11ytable object to print.}

\item{...}{Other arguments to pass.}
}
\description{
Print an a11ytable object to see information about the
sheet content.
}
\examples{
x <- as_a11ytable(lfs_tables)
print(x)

}
