% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11ytable.R
\name{tbl_sum.a11ytable}
\alias{tbl_sum.a11ytable}
\title{Provide A Succinct Summary Of An 'a11ytable' Object}
\usage{
\method{tbl_sum}{a11ytable}(x, ...)
}
\arguments{
\item{x}{An a11ytable-class object to summarise.}

\item{...}{Other arguments to pass.}
}
\value{
Named character vector.
}
\description{
A brief text description of an a11ytable-class object.
}
\examples{
# Print with description
print(demo_a11ytable)

# Print description only (package 'tibble' must be installed)
tibble::tbl_sum(demo_a11ytable)

}
