% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11ytable.R
\name{as_a11ytable}
\alias{as_a11ytable}
\alias{is_a11ytable}
\title{Coerce To An 'a11ytable' Object}
\usage{
as_a11ytable(x)

is_a11ytable(x)
}
\arguments{
\item{x}{A data.frame object to coerce.}
}
\value{
\code{as_a11ytable} returns an object of class a11ytable if possible.
\code{is_a11ytable} returns \code{TRUE} if the object has class
a11ytable, otherwise \code{FALSE}.
}
\description{
Functions to check if an object is an a11ytable, or coerce it if possible.
}
\examples{
# Create an a11ytable with in-built demo dataframe, mtcars_df2. We can use
# 'as_a11ytable' rather than 'create_a11ytable' because the data is already
# in the right format.
x <- as_a11ytable(mtcars_df2)

# Test the object's class
is_a11ytable(x)

}
