% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_a11y_wb}
\alias{create_a11y_wb}
\title{Create An Accessible Workbook}
\usage{
create_a11y_wb(content)
}
\arguments{
\item{content}{A data.frame-class object containing your data.}
}
\value{
A Workbook-class object.
}
\description{
Given a correctly-formatted dataframe input, create a workbook object with
the required tab content (cover, contents, notes (optional) and one tab
per table). A shortcut for building workbooks with the add_*() functions.
}
\details{
The 'content' object should adhere to the form demonstrated in
\code{\link{lfs_tables}}
}
\examples{
\dontrun{ create_a11y_wb(content) }
}
