% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_a11y_wb}
\alias{create_a11y_wb}
\title{Create An Accessible Workbook}
\usage{
create_a11y_wb(content)
}
\arguments{
\item{content}{An a11ytable-class object containing the data and information
needed to create your workbook.}
}
\value{
A Workbook-class object.
}
\description{
Populate an {openxlsx} workbook object with content from a a11ytable-class
input.
}
\details{
You can create an a11ytable object with
\code{\link{new_a11ytable}} (or \code{\link{as_a11ytable}}).
}
\examples{
x <- as_a11ytable(lfs_tables)
create_a11y_wb(x)

}
