% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11ytable.R
\name{create_a11ytable}
\alias{create_a11ytable}
\title{Create An 'a11ytable' Object}
\usage{
create_a11ytable(
  tab_titles,
  sheet_types = c("cover", "contents", "notes", "tables"),
  sheet_titles,
  blank_cells = NA_character_,
  sources = NA_character_,
  tables
)
}
\arguments{
\item{tab_titles}{Required character vector, one value per sheet. Each title
will appear on each tab of the final spreadsheet output. Keep brief.
For example: 'Cover', 'Contents', 'Notes', 'Table 1'.}

\item{sheet_types}{Required character vector, one value per sheet. Sheets
that don't contain publication tables ('meta' sheets) should be of type
'contents', 'cover' or 'notes'. Sheets that contain statistical tables of
data are type 'tables'.}

\item{sheet_titles}{Required character vector, one value per sheet. The main
title for each sheet, which will appear in cell A1 (top-left corner).}

\item{blank_cells}{Optional character vector, one value per sheet. A short
sentence to explain the reason for any blank cells in the sheet. Most
likely to be used with sheet type 'tables'.}

\item{sources}{Optional character vector, one value per sheet. The origin of the data for a given sheet. Supply as
\code{NA_character_} if empty. To be used with sheet type 'tables'.}

\item{tables}{Required list of data.frames, one per sheet. See details.}
}
\value{
An a11ytable-class object.
}
\description{
Create a new a11ytable-class object, which is a data.frame that contains all
the information needed in your publication. In turn, this will be used to
populate an 'openxlsx' Workbook-class object with the function
\code{\link{generate_workbook}}.
}
\details{
Formats for data.frames in the 'tables' argument, depending on the
sheet type.
\itemize{
\item Sheet type 'cover': one row per subsection, with columns for
'Subsection title' and 'Subsection text'. For example, a section with
contact details might have 'Contact details' as the subsection title
and a telephone number and email address in the body-text column. Use
linebreaks (i.e. '\\n') to put multiple rows in the same body text.
Don't break information into separate spreadsheet rows if they belong
in the same subsection).
\item Sheet type 'contents': one row per sheet, two columns suggested at
least ('Tab title' and 'Worksheet title').
\item Sheet type 'notes': one row per note, two columns suggested ('Note
number', 'Note text').
\item Sheet type 'tables': a tidy, rectangular data.frame containing the
data to be published.
}
}
\examples{
\dontrun{
x <- create_a11ytable(
    tab_titles   = mtcars_df$tab_title,
    sheet_types  = mtcars_df$sheet_type,
    sheet_titles = mtcars_df$sheet_title,
    blank_cells  = mtcars_df$blank_cells,
    sources      = mtcars_df$source,
)
is_a11ytable(x)
}

}
