% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11ytable.R
\name{create_a11ytable}
\alias{create_a11ytable}
\title{Create An 'a11ytable' Object}
\usage{
create_a11ytable(
  tab_titles,
  sheet_types = c("cover", "contents", "notes", "tables"),
  sheet_titles,
  blank_cells = NA_character_,
  sources = NA_character_,
  custom_rows = list(NA_character_),
  tables
)
}
\arguments{
\item{tab_titles}{Required character vector, one value per sheet. Each title
will appear literally on each tab of the final spreadsheet output. Keep
brief. Letters and numbers only; use underscores for spaces. For example:
'Cover', 'Contents', 'Notes', 'Table_1'. Will be corrected automatically
if non-conforming.}

\item{sheet_types}{Required character vector, one value per sheet. Sheets
that don't contain publication tables ('meta' sheets) should be of type
'contents', 'cover' or 'notes'. Sheets that contain statistical tables of
data are type 'tables'.}

\item{sheet_titles}{Required character vector, one value per sheet. The main
title for each sheet, which will appear in cell A1 (top-left corner).}

\item{blank_cells}{Optional character vector, one value per sheet. A short
sentence to explain the reason for any blank cells in the sheet. Supply
as \code{NA_character_} if empty. Most likely to be used with sheet type
'tables'.}

\item{sources}{Optional character vector, one value per sheet. The origin of
the data for a given sheet. Supply as \code{NA_character_} if empty. To
be used with sheet type 'tables'.}

\item{custom_rows}{Optional list of character vectors. One list element per
sheet, one character vector element per row of pre-table metadata. Supply
as \code{NA_character_} if empty. To be used with sheet type 'tables',
but can also be used for sheet types 'contents' and 'notes'.}

\item{tables}{Required list of data.frames (though the cover sheet may be
supplied as a list), one per sheet. See details.}
}
\value{
An object with classes 'a11ytable', 'tbl' and 'data.frame'.
}
\description{
Create a new a11ytable-class object, which is a dataframe that contains all
the information needed in your output spreadsheet. In turn, the object
created by  this function can be used to populate an 'openxlsx'
Workbook-class object with the function \code{\link{generate_workbook}}.
}
\details{
\subsection{How to supply data to the 'tables' argument}{

Formats for the elements collected as a list and passed to the 'tables'
argument, depending on the sheet type.

\itemize{
\item Sheet type 'cover': either (a) a list where each element name is a
section header and each element's content is a character vector
whose elements will make up separate rows of that section
(recommended), or (b) a data.frame with one row per subsection, with
one column for section titles and one column for corresponding for
that section's body text. For example, you may have a section with
the title 'Contact details' that contains an email address and
telephone number. You can use linebreaks (i.e. '\\n') to separate
text into paragraphs.
\item Sheet type 'contents': one row per sheet, two columns suggested at
least ('Tab title' and 'Worksheet title').
\item Sheet type 'notes': one row per note, two columns suggested ('Note
number', 'Note text'), where notes are in the form '[note 1]'.
\item Sheet type 'tables': a tidy, rectangular data.frame containing the
data to be published. It's the user's responsibility to add notes in
the form '[note 1]' to column headers, or in a special 'Notes' row.
}
}

\subsection{How to supply hyperlinks}{

You can provide text in Markdown link syntax (e.g.
'[GOV.UK](https://www.gov.uk)', adding 'mailto:' before an
email address) and the containing cell will be rendered as a hyperlink in the
output spreadsheet. Note that whole cells will become hyperlinks; there is no
support for selected words in a sentence to be rendered as a hyperlink.

Hyperlinks can be supplied in two locations:

\itemize{
\item To the 'tables' argument for sheet type 'cover' only. It's
recommended to supply the cover information as a list rather than a
data.frame, which will allow you to make specific rows within a
section into hyperlinks. For example, in a 'Contact us' section you
might want a row containing some preamble (no hyperlink), a cell
containing a phone number (no hyperlink) and a cell containing an
email address (hyperlinked).
\item To the 'source' argument for sheets that contain tables of data.
}
}
}
\examples{
# Create an a11ytable with in-built demo dataframe, mtcars_df3
x <- create_a11ytable(
  tab_titles   = mtcars_df3$tab_title,
  sheet_types  = mtcars_df3$sheet_type,
  sheet_titles = mtcars_df3$sheet_title,
  blank_cells  = mtcars_df3$blank_cells,
  sources      = mtcars_df3$source,
  custom_rows  = mtcars_df3$custom_rows
  tables       = mtcars_df3$table
)

# Test that 'a11ytable' is one of the object's classes
is_a11ytable(x)

}
