% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_a11ytable}
\alias{new_a11ytable}
\title{Create An a11ytable-class Object}
\usage{
new_a11ytable(
  tab_titles,
  sheet_types = c("cover", "contents", "notes", "tables"),
  sheet_titles,
  sources = NULL,
  table_names,
  tables
)
}
\arguments{
\item{tab_titles}{Required character vector, one value per sheet to be
created. Each title will appear on each tab of the final spreadsheet
output. Keep brief.}

\item{sheet_types}{Required character vector, one value per sheet to be
created. Sheets that don't contain publication tables ('meta' sheets)
should be of type 'contents', 'cover' or 'notes'. Sheets that contain
statistical tables of data are type 'tables'.}

\item{sheet_titles}{Required character vector, one value per sheet to be
created. The main title of the sheet, which will appear in cell A1
(top-left corner) of each sheet.}

\item{sources}{Optional character vector, one value per sheet to be created.
The origin of the data for a given sheet. Supply as \code{NA_character_}
if empty.}

\item{table_names}{Required character vector, one value per table to be
created (i.e. one per sheet). A short, descriptive, identifying name to
give the 'named range' of cells that compose the table in the final
spreadsheet output. Makes sheet navigation more accessible.}

\item{tables}{Required list of data.frames, one per sheet. See details.}
}
\value{
An a11ytables-class object.
}
\description{
Create a new a11ytable-class object, which is a data.frame that contains all
the information needed in your publication. In turn, this will be used to
populate an 'openxlsx' Workbook-class object with the function
\code{\link{create_a11y_wb}}.
}
\details{
Formats for data.frames in the 'tables' argument, depending on the
sheet type.
\itemize{
\item Sheet type 'cover': one row per subsection, with columns for
'Subsection title' and 'Subsection text'. For example, a section with
contact details might have 'Contact details' as the subsection title
and a telephone number and email address in the body-text column. Use
linebreaks (i.e. '\n') to put multiple rows in the same body text.
Don't break information into separate spreadsheet rows if they belong
in the same subsection).
\item Sheet type 'contents': one row per sheet, two columns suggested at
least ('Tab title' and 'Worksheet title').
\item Sheet type 'notes': one row per note, two columns suggested ('Note
number', 'Note text').
\item Sheet type 'tables': a tidy, rectangular data.frame containing the
data to be published.
}
}
\examples{
\dontrun{
x <- new_a11ytable(
    tab_titles   = lfs_tables$tab_title,
    sheet_types  = lfs_tables$sheet_type,
    sheet_titles = lfs_tables$sheet_title,
    sources      = lfs_tables$source,
    table_names  = lfs_tables$table_name,
)
is_a11ytable(x)
}

}
