% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11ytable.R
\name{summary.a11ytable}
\alias{summary.a11ytable}
\title{Summarise An 'a11ytable' Object}
\usage{
\method{summary}{a11ytable}(object, ...)
}
\arguments{
\item{object}{An a11ytable-class object for which to get a summary.}

\item{...}{Other arguments to pass.}
}
\description{
A concise result summary of an a11ytable-class object to see information about
the sheet content.
}
\examples{
# Coerce a compliant data.frame to a11ytables-class
x <- as_a11ytable(demo_df)

# Print a summary of the a11ytable-class object
summary(x)

# Alternatively, look at the structure
str(x, max.level = 2)

}
