% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{summary.a11ytable}
\alias{summary.a11ytable}
\title{Summarise An 'a11ytable' Object}
\usage{
\method{summary}{a11ytable}(x, ...)
}
\arguments{
\item{...}{Other arguments to pass.}

\item{object}{An a11ytable-class object to get a summary for.}
}
\description{
A concise result summary of an a11ytable-class object to see information about
the sheet content.
}
\examples{
\dontrun{
x <- as_a11ytable(mtcars_df)
summary(x)
}

}
