% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11ytable.R
\name{summary.a11ytable}
\alias{summary.a11ytable}
\title{Summarise An 'a11ytable' Object}
\usage{
\method{summary}{a11ytable}(object, ...)
}
\arguments{
\item{object}{An a11ytable-class object for which to get a summary.}

\item{...}{Other arguments to pass.}
}
\description{
A concise result summary of an a11ytable-class object to see information about
the sheet content.
}
\examples{
# Create an a11ytable with in-built demo dataframe, mtcars_df3. We can use
# 'as_a11ytable' rather than 'create_a11ytable' because the data is already
# in the right format.
x <- as_a11ytable(mtcars_df3)

# Print summary of a11ytable-class object
summary(x)

}
