% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11ytable.R
\name{tbl_sum.a11ytable}
\alias{tbl_sum.a11ytable}
\title{Provide A Succinct Summary Of An 'a11ytable' Object}
\usage{
\method{tbl_sum}{a11ytable}(x, ...)
}
\arguments{
\item{x}{An a11ytable-class object to summarise.}

\item{...}{Other arguments to pass.}
}
\value{
Named character vector.
}
\description{
A brief textual description of an a11ytable-class object.
}
\examples{
\dontrun{
# Create an a11ytable with in-built demo dataframe, mtcars_df
x <- as_a11ytable(mtcars_df)

# Print description only
tbl_sum(x)

# Print with description
print(x)
}

}
