% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11ytable.R
\name{tbl_sum.a11ytable}
\alias{tbl_sum.a11ytable}
\title{Provide A Succinct Summary Of An 'a11ytable' Object}
\usage{
\method{tbl_sum}{a11ytable}(x, ...)
}
\arguments{
\item{x}{An a11ytable-class object to summarise.}

\item{...}{Other arguments to pass.}
}
\value{
Named character vector.
}
\description{
A brief textual description of an a11ytable-class object.
}
\examples{
# Create an a11ytable with in-built demo dataframe, mtcars_df3. We can use
# 'as_a11ytable' rather than 'create_a11ytable' because the data is already
# in the right format.
x <- as_a11ytable(mtcars_df3)

# Print description only
tibble::tbl_sum(x)

# Print with description
print(x)

}
